% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-add-info.R
\name{add_nTaxaInRank}
\alias{add_nTaxaInRank}
\title{Count the number of taxa per taxonomic rank in each sample}
\usage{
add_nTaxaInRank(report)
}
\arguments{
\item{report}{A report generated by mergeReports() or loaded with load_STDreports().}
}
\value{
An updated version of the report provided as input, with a new column containing
the number of taxa present per taxonomic rank in each sample.
}
\description{
Count the number of taxa per taxonomic rank in each sample
}
