% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-add-info.R
\name{addMinimiserData}
\alias{addMinimiserData}
\title{Add minimiser data to the report}
\usage{
addMinimiserData(report, ref_db)
}
\arguments{
\item{report}{A report generated by mergeReports() or loaded with load_STDreports().}

\item{ref_db}{A Kraken2 reference database 'inspect.txt' file loaded with loadReference().}
}
\value{
An updated version of the report provided as input, with new columns containing
the number of taxon-level and clade-level minimisers from the database available for each
taxon.
}
\description{
Add minimiser data to the report
}
