% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-add-info.R
\name{addRank}
\alias{addRank}
\title{Define rank for a given taxon hierarchy from an MPA-style report}
\usage{
addRank(taxon)
}
\arguments{
\item{taxon}{A taxon hierarchy from an MPA-style report.}
}
\value{
The rank corresponding to the given taxon hierarchy.
}
\description{
Define rank for a given taxon hierarchy from an MPA-style report
}
