% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-add-info.R
\name{addSampleSize}
\alias{addSampleSize}
\title{CALCULATE SAMPLE SIZES}
\usage{
addSampleSize(report)
}
\arguments{
\item{report}{A report.}
}
\value{
An updated version of the input dataframe, with a new column containing sample sizes.
}
\description{
CALCULATE SAMPLE SIZES
}
