% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_barplot}
\alias{adjust_barplot}
\title{Adjust ggplot2 bar plot.}
\usage{
adjust_barplot(plot, n_samples, include_sample_names, orientation, filename)
}
\arguments{
\item{plot}{A ggplot2 object.}

\item{n_samples}{The number of samples that will be displayed in the plot.}

\item{include_sample_names}{Whether sample names should be included in the plot.}

\item{orientation}{The orientation of the plot ("horizontal"/"h" or "vertical"/"v").}

\item{filename}{The file name for the plot.}
}
\value{
An updated version of the plot, with the appropriate tick names.
}
\description{
Adjust ggplot2 bar plot.
}
