% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-assess-statistics.R
\name{assessMinimiserRatio}
\alias{assessMinimiserRatio}
\title{Calculate the proportion of minimisers of each taxon found in each sample}
\usage{
assessMinimiserRatio(report)
}
\arguments{
\item{report}{A report generated by mergeReports() or loaded with load_STDreports().
Note that addMinimiserData() must be run before running assessMinimiserRatio().}
}
\value{
An updated version of the report provided, containing new columns with the
proportions of taxon-level and clade-level minimisers for each taxon in each sample.
}
\description{
Calculate the proportion of minimisers of each taxon found in each sample
}
