% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-assess-statistics.R
\name{assessStatistics}
\alias{assessStatistics}
\title{Assess the statistical significance of the Kraken2 results}
\usage{
assessStatistics(report, ref_db)
}
\arguments{
\item{report}{A report generated by mergeReports() and processed with addSampleSize(), addMinimiserData()
and assessMinimiserRatio().}

\item{ref_db}{A Kraken2 reference database 'inspect.txt' file loaded with loadReference().}
}
\value{
An updated version of the report provided as input, with new columns containing statistical
significance results.
}
\description{
Assess the statistical significance of the Kraken2 results
}
