% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-assess-statistics.R
\name{calculate_p_value}
\alias{calculate_p_value}
\title{Calculate the p-values for each taxon in each sample}
\usage{
calculate_p_value(
  sample_n_uniq_minimisers_taxon,
  db_n_minimisers_taxon,
  sample_size,
  ref_db
)
}
\arguments{
\item{sample_n_uniq_minimisers_taxon}{The number of unique minimisers found for a given taxon in a
given sample.}

\item{db_n_minimisers_taxon}{The total number of clade-level minimisers available for a given taxon in
the Kraken2 reference database.}

\item{sample_size}{The total number of reads analysed by Kraken2 for each sample (or sample size).}

\item{ref_db}{The Kraken2 reference database 'inspect.txt' file loaded with loadReference().}
}
\description{
Calculate the p-values for each taxon in each sample
}
