% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_columns}
\alias{check_columns}
\title{Check if a set of column names is present in a given dataframe}
\usage{
check_columns(df, columns)
}
\arguments{
\item{df}{Dataframe where given column names will be looked for.}

\item{columns}{A vector with column names that are expected to be present
in the given dataframe.}
}
\description{
This function takes a dataframe and a vector containing column names that
are expected to be found in the given dataframe; then it checks if each
column is indeed present, throwing an error otherwise.
}
\examples{
check_columns(df = my_dataframe, columns = c("col1", "col2", "col3"))

}
