% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_directory}
\alias{check_directory}
\title{CARRY OUT CHECKS ON A GIVEN DIRECTORY}
\usage{
check_directory(dirpath)
}
\arguments{
\item{dirpath}{Path to a given directory.}
}
\value{
Checked path to the given directory.
}
\description{
This function takes the path to a given directory and performs a few checks:
\enumerate{
\item First, it checks if the directory exists.
\item Second, it checks if the directory is empty.
\item Last, it ensures that the last character in the path is a slash symbol ("/").
}
}
\examples{
check_directory(dirpath = "path/to/directory")

}
\seealso{
\code{\link[=check_report_directory]{check_report_directory()}} for another function that carries out
checks based on a given directory path.
}
