% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_domain}
\alias{check_domain}
\title{Check if the domain specified is valid}
\usage{
check_domain(domain)
}
\arguments{
\item{domain}{A domain ("Viruses", "Bacteria", "Archaea", or "Eukaryota").}
}
\description{
This function takes a domain and checks if it is valid, i.e. one of "Viruses",
"Bacteria", "Archaea", or "Eukaryota".
}
