% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_file}
\alias{check_file}
\title{CARRY OUT CHECKS ON A GIVEN FILE}
\usage{
check_file(file_path)
}
\arguments{
\item{file_path}{Path to a given file.}
}
\description{
This function takes the path to a given file and performs a few checks:
\enumerate{
\item First, it checks if the file exists.
\item Second, it checks if the file is empty.
}
}
\examples{
check_file(file_path = "path/to/file")

}
