% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{check_for_empty_files}
\alias{check_for_empty_files}
\title{Exclude empty reports from the analysis}
\usage{
check_for_empty_files(file_list)
}
\arguments{
\item{file_list}{A list containing paths to standard or MPA-style reports.}
}
\value{
An updated version of file list, without empty reports.
}
\description{
Exclude empty reports from the analysis
}
