% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{check_mpa_lines}
\alias{check_mpa_lines}
\title{Ensure taxon hierarchies in an MPA-style report are "complete"}
\usage{
check_mpa_lines(mpa_reports)
}
\arguments{
\item{mpa_reports}{A report loaded by load_MPAreports().}
}
\value{
An updated version of the report provided, with all taxon
hierarchies being "complete".
}
\description{
Ensure taxon hierarchies in an MPA-style report are "complete"
}
