% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_organism}
\alias{check_organism}
\title{Check if the organism specified is present in the reports}
\usage{
check_organism(std_reports_path, mpa_reports_path, organism)
}
\arguments{
\item{organism}{The name of the organism being analysed (species-level). For
example, if your input data for Kraken2 comes from human samples, then the
organism should be "Homo sapiens".}
}
\description{
This function takes an organism at the species level (e.g. "Homo sapiens")
and checks if it is present in the standard and MPA-style reports.
}
\examples{
check_organism(organism = "Homo sapiens")

}
