% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_prefix}
\alias{check_prefix}
\title{CARRY OUT CHECKS ON A PREFIX}
\usage{
check_prefix(prefix)
}
\arguments{
\item{prefix}{A given prefix that will later be added to output file names.}
}
\value{
Checked prefix.
}
\description{
This function takes a prefix and performs a few checks:
\enumerate{
\item First, it checks if the prefix corresponds to the appropriate class ("character").
\item Second, it ensures that there is an underscore symbol ("_") at the end of the prefix.
}
}
\examples{
check_prefix(prefix = "prefix")

}
