% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_report_directory}
\alias{check_report_directory}
\title{Carry out checks on directories that contain standard or MPA-style reports}
\usage{
check_report_directory(dirpath, report_format)
}
\arguments{
\item{dirpath}{Path to a given directory.}

\item{report_format}{One of 'std' (standard) or 'mpa' (MPA-style), indicating
the report format that is expected in the directory specified.}
}
\value{
Checked path to the given directory.
}
\description{
This function takes the path to a directory containing either standard or
MPA-style reports and performs a few checks:
\itemize{
\item If the directory is expected to contain MPA-style reports, the function will check
whether MPA-style reports are indeed present.
\item If the directory is expected to contain standard reports, the function will check
whether standard reports are indeed present.
\item The function also ensures that the directory provided has only one report format,
throwing an error otherwise.
}
}
\examples{
check_report_directory(dirpath = "path/to/directory", report_format = "std")

}
\seealso{
\code{\link[=check_directory]{check_directory()}} for another function that carries out
checks based on a given directory path.
}
