% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_pdf_height}
\alias{determine_pdf_height}
\title{DETERMINE HEIGHT FOR PDF FILE}
\usage{
determine_pdf_height(n_elements, base_size = 2, factor = 1)
}
\arguments{
\item{n_elements}{Number of elements on the y-axis.}

\item{base_size}{Base plot height when there is only one element on the y-axis.}

\item{factor}{Value to help adjust the height.}
}
\value{
Plot height for PDF file.
}
\description{
This function takes a number of elements (e.g. number of genes) and determines the height
that should be used for a plot to be saved in PDF format. This function is tailored for the
plots generated by plot_mahalanobis_with_metadata() and plot_mahalanobis_with_driver_genes(),
so should be used with caution for any other ends.
}
