% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_pdf_width}
\alias{determine_pdf_width}
\title{DETERMINE WIDTH FOR PDF FILE}
\usage{
determine_pdf_width(n_elements, base_size = 1, factor = 1)
}
\arguments{
\item{base_size}{Base plot width when there is only one element on the x-axis.}

\item{factor}{Value to help adjust the width.}
}
\value{
Plot width for PDF file.
}
\description{
This function takes a number of elements (e.g. number of samples) and determines the width
that should be used for a plot to be saved in PDF format. This function is tailored for the
plots generated by plot_mahalanobis_with_metadata() and plot_mahalanobis_with_driver_genes(),
so should be used with caution for any other ends.
}
\details{
@param n_elements Number of elements on the x-axis.
}
