% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportPlot}
\alias{exportPlot}
\title{Save plot to output file}
\usage{
exportPlot(plot, filename, outdir, fig_width, fig_height)
}
\arguments{
\item{plot}{A ggplot2 object.}

\item{filename}{The file name for the plot.}

\item{outdir}{Output directory where the plot should be saved.}

\item{fig_width}{The width of the output PDF file.}

\item{fig_height}{The height of the output PDF file.}
}
\description{
Save plot to output file
}
