% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-assess-statistics.R
\name{get_binomial_parameters}
\alias{get_binomial_parameters}
\title{Calculate the parameters of a binomial distribution}
\usage{
get_binomial_parameters(db_n_minimisers_taxon, sample_size, ref_db)
}
\arguments{
\item{db_n_minimisers_taxon}{The total number of clade-level minimisers available for a given taxon in
the Kraken2 reference database.}

\item{sample_size}{The total number of reads analysed by Kraken2 for each sample (or sample size).}

\item{ref_db}{The Kraken2 reference database 'inspect.txt' file loaded with loadReference().}
}
\description{
Calculate the parameters of a binomial distribution
}
