% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nDomainReads}
\alias{get_nDomainReads}
\title{Number of reads under each domain for each sample.}
\usage{
get_nDomainReads(report, include_eukaryotes)
}
\description{
Number of reads under each domain for each sample.
}
