% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-classification-summary.R
\name{getClassificationProportion}
\alias{getClassificationProportion}
\title{Get the proportion of classified/unclassified reads per sample}
\usage{
getClassificationProportion(report)
}
\arguments{
\item{report}{A report generated by mergeReports().}
}
\value{
A summary dataframe with the proportion of classified/unclassified reads per sample.
}
\description{
Get the proportion of classified/unclassified reads per sample
}
