% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-classification-summary.R
\name{getClassificationSummary}
\alias{getClassificationSummary}
\title{Get the numbers of classified/unclassified reads per sample}
\usage{
getClassificationSummary(report)
}
\arguments{
\item{report}{A report generated by mergeReports().}
}
\value{
A summary dataframe with the numbers of classified/unclassified reads per sample.
}
\description{
Get the numbers of classified/unclassified reads per sample
}
