% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-significance.R
\name{getSignificanceSummary}
\alias{getSignificanceSummary}
\title{Prepare report for plotSignficanceSummary()}
\usage{
getSignificanceSummary(report)
}
\arguments{
\item{report}{A report generated by mergeReports() and processed by
assessStatistics().}
}
\value{
An updated version of the report provided as input, suitable for
use by the function plotSignficanceSummary().
}
\description{
Prepare report for plotSignficanceSummary()
}
