% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handlePlot}
\alias{handlePlot}
\title{Handle plot based on the user-defined arguments}
\usage{
handlePlot(plot, filename, prefix, outdir, fig_width, fig_height, return_plot)
}
\arguments{
\item{plot}{A ggplot2 object.}

\item{filename}{The file name for the plot.}

\item{prefix}{The prefix that should be added to the file name.}

\item{outdir}{Output directory where the plot should be saved.}

\item{fig_width}{The width of the output PDF file.}

\item{fig_height}{The height of the output PDF file.}

\item{return_plot}{Whether plot should be returned.}
}
\value{
Depending on the arguments provided, the plot will be returned.
}
\description{
Handle plot based on the user-defined arguments
}
