% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handlePrefix}
\alias{handlePrefix}
\title{Add prefix to output file name}
\usage{
handlePrefix(filename, prefix)
}
\arguments{
\item{filename}{The file name for the plot.}

\item{prefix}{The prefix that should be added to the file name.}
}
\value{
An updated version of the file name provided.
}
\description{
Add prefix to output file name
}
