% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_MPAreports}
\alias{load_MPAreports}
\title{LOAD MPA-STYLE REPORTS}
\usage{
load_MPAreports(mpa_reports_dir, samples_to_remove, verbose)
}
\arguments{
\item{mpa_reports_dir}{Path to a directory containing MPA-style reports.}

\item{verbose}{Whether to output log messages.}
}
\value{
A dataframe (tibble) with the content of all MPA-style reports from the specified directory.
}
\description{
This function takes a path to a directory containing MPA-style reports, reading and processing
all reports into a single dataframe.
}
