% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_STDreports}
\alias{load_STDreports}
\title{Load standard reports}
\usage{
load_STDreports(std_reports_dir, samples_to_remove)
}
\arguments{
\item{std_reports_dir}{Path to a directory containing standard reports.}

\item{samples_to_remove}{A list of samples to be removed, loaded with loadSamplesToRemove().}
}
\value{
A dataframe (tibble) with the content of all standard reports from the specified directory.
}
\description{
This function takes a path to a directory containing standard reports, reading and collating
all sample-level reports into a single dataframe.
}
