% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadMetadata}
\alias{loadMetadata}
\title{Load a metadata table}
\usage{
loadMetadata(metadata)
}
\arguments{
\item{metadata}{Path to a metadata table.}
}
\value{
A dataframe containing the metadata.
}
\description{
This function takes the path to a metadata file and reads the metadata table using
readr::read_delim().
}
