% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadReference}
\alias{loadReference}
\title{Load the 'inspect.txt' file from a Kraken2 reference database}
\usage{
loadReference(reference_path)
}
\arguments{
\item{reference_path}{Path to an 'inspect.txt' file inside a Kraken2 reference database.}
}
\value{
A dataframe containing the information from the 'inspect.txt' file.
}
\description{
This function takes the path to an 'inspect.txt' file inside a Kraken2 reference database
and reads the table. The first 7 lines are skipped, as they correspond to header lines.
Column names are added before the reference database dataframe is returned.
}
