% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadSamplesToRemove}
\alias{loadSamplesToRemove}
\title{Load list of samples to be removed from the SPARKI analysis}
\usage{
loadSamplesToRemove(filepath)
}
\arguments{
\item{filepath}{Path to a tab-delimited file in which each line is a sample ID.}
}
\value{
A list containing the samples IDs from the input file.
}
\description{
This function takes the path to a tab-delimited file, in which each line should
be a sample ID, and returns a list with these IDs. Note that the sample IDs need
to match those that are present in the reports to be loaded with load_STDreports()
and load_MPAreports().
}
