% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-merge.R
\name{mergeReports}
\alias{mergeReports}
\title{MERGE STANDARD AND MPA-STYLE REPORTS}
\usage{
mergeReports(std_reports, mpa_reports, samples_to_remove)
}
\arguments{
\item{std_reports}{A dataframe (tibble) with standard reports, loaded with load_STDreports().}

\item{mpa_reports}{A dataframe (tibble) with MPA-style reports, loaded with load_MPAreports().}
}
\value{
An updated version of the dataframe (tibble) with standard reports, now containing the
information from the dataframe (tibble) with MPA-style reports.
}
\description{
This function takes a standard and an MPA-style dataframe, loaded with load_STDreports() and
load_MPAreports() respectively, and merges the latter into the former.
}
