% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_delimited_list}
\alias{parse_delimited_list}
\title{PARSE SYMBOL-DELIMITED LIST TO GET INDIVIDUAL ELEMENTS}
\usage{
parse_delimited_list(del_list, delimiter)
}
\value{
A vector with individual elements.
}
\description{
This function takes a symbol-delimited list of elements and splits it up into individual elements.
Symbols can be anything, e.g. ",", ";", "/", "//", "@", "." etc.
}
\details{
@param del_list A list with symbol-delimited values; the symbol can be a comma, for example.
}
