% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-classification-summary.R
\name{plotClassificationProportion}
\alias{plotClassificationProportion}
\title{Plot the proportion of classified reads across samples}
\usage{
plotClassificationProportion(report, return_plot, outdir, prefix = "")
}
\arguments{
\item{report}{A report generated by mergeReports() or loaded with load_STDreports().}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where plot should be saved.}

\item{prefix}{Prefix to be added to output file name.}
}
\value{
If return_plot = TRUE, returns a bar plot with the classification proportion.
}
\description{
Plot the proportion of classified reads across samples
}
