% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-classification-summary.R
\name{plotClassificationSummary_violin}
\alias{plotClassificationSummary_violin}
\title{PLOT NUMBERS OF CLASSIFIED/UNCLASSIFIED READS PER SAMPLE (VIOLIN PLOT)}
\usage{
plotClassificationSummary_violin(report, return_plot, outdir, prefix = "")
}
\arguments{
\item{report}{Standard Kraken2 report.}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}
}
\description{
This function takes a standard Kraken2 report and creates a violin plot showing
the number of classified and unclassified reads per sample.
}
