% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-domain-reads.R
\name{plotDomainReads_barplot}
\alias{plotDomainReads_barplot}
\title{PLOT NUMBERS OF CLASSIFIED READS PER DOMAIN (BAR PLOT)}
\usage{
plotDomainReads_barplot(
  report,
  include_sample_names = FALSE,
  orientation = "vertical",
  include_eukaryotes = FALSE,
  return_plot,
  outdir,
  prefix = ""
)
}
\arguments{
\item{report}{Kraken2 report, either in standard or MPA-style format.}

\item{include_sample_names}{Whether sample names should be displayed.}

\item{orientation}{Whether plot should be horizontally or vertically oriented.}

\item{include_eukaryotes}{Whether to include eukaryotes in the plot.}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}

\item{prefix}{Prefix to be added to output plot name.}
}
\description{
This function takes a Kraken2 report, either in standard or MPA-style format, and creates
a bar plot showing the number of classified reads per domain.
}
