% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-minimisers.R
\name{plotMinimisers_dotplot}
\alias{plotMinimisers_dotplot}
\title{Plot proportion of minimisers and significance per sample}
\usage{
plotMinimisers_dotplot(
  report,
  domain,
  fig_width,
  fig_height,
  return_plot,
  outdir,
  prefix
)
}
\arguments{
\item{report}{A report generated by mergeReports() and processed by
assessStatistics().}

\item{domain}{Domain of interest.}

\item{fig_width}{Width for output PDF file.}

\item{fig_height}{Height for output PDF file.}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}

\item{prefix}{Prefix to be added to output file name.}
}
\value{
If return_plot = TRUE, returns a dot plot with the minimiser
proportions and adjusted p-values.
}
\description{
Plot proportion of minimisers and significance per sample
}
