% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-significance.R
\name{plotSignificanceSummary}
\alias{plotSignificanceSummary}
\title{Plot a summary of statistical significance results}
\usage{
plotSignificanceSummary(report, return_plot, outdir, prefix)
}
\arguments{
\item{report}{A report generated by mergeReports() and processed
with assessStatistics().}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}

\item{prefix}{Prefix to be added to output file name.}
}
\value{
If return_plot = TRUE, returns the plot.
}
\description{
Plot a summary of statistical significance results
}
