% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prepare_data}
\alias{prepare_data}
\title{PREPARE DATA}
\usage{
prepare_data(
  std_reports_path,
  mpa_reports_path,
  reference_path,
  metadata_path,
  metadata_sample_col,
  metadata_columns,
  outdir_path,
  prefix,
  verbose,
  domain,
  remove
)
}
\arguments{
\item{std_reports_path}{Path to directory containing standard Kraken2 reports.}

\item{mpa_reports_path}{Path to directory containing MPA-style reports.}

\item{reference_path}{Path to file containing information on a Kraken2 reference database.}

\item{metadata_path}{Path to metadata file.}

\item{verbose}{Verbosity level.}
}
\value{
Processed data ready for downstream steps.
}
\description{
PREPARE DATA
}
