% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-read-distribution.R
\name{prepare_for_plotDistribution}
\alias{prepare_for_plotDistribution}
\title{Prepare report for plotDistribution_histogram() and plotDistribution_violin()}
\usage{
prepare_for_plotDistribution(report)
}
\arguments{
\item{report}{A report generated by mergeReports().}
}
\value{
An updated version of the report provided as input, suitable for
use by the functions plotDistribution_histogram() and plotDistribution_violin().
}
\description{
Prepare report for plotDistribution_histogram() and plotDistribution_violin()
}
