% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-domain-reads.R
\name{prepare_for_plotDomainReads}
\alias{prepare_for_plotDomainReads}
\title{Get the numbers of classified reads per domain per sample}
\usage{
prepare_for_plotDomainReads(report, include_eukaryotes)
}
\arguments{
\item{report}{A report generated by mergeReports().}

\item{include_eukaryotes}{Whether eukaryotes should be included.}
}
\value{
A summary dataframe with the numbers of classified reads per
domain per sample.
}
\description{
Get the numbers of classified reads per domain per sample
}
