% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-minimisers.R
\name{prepare_for_plotMinimisers}
\alias{prepare_for_plotMinimisers}
\title{Prepare report for plotMinimisers_dotplot()}
\usage{
prepare_for_plotMinimisers(report, domain_of_interest)
}
\arguments{
\item{report}{A report generated by mergeReports() and with statistical
assessment results generated by assessStatistics().}

\item{domain_of_interest}{The domain that should be included in the plot
generated by plotMinimisers_dotplot().}
}
\value{
An updated version of the report provided as input, suitable for
use by the function plotMinimisers_dotplot().
}
\description{
Prepare report for plotMinimisers_dotplot()
}
