% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_barplot_orientation}
\alias{process_barplot_orientation}
\title{Process a ggplot2 bar plot to define its orientation and dimensions}
\usage{
process_barplot_orientation(plot, n_samples, orientation, factor)
}
\arguments{
\item{plot}{A ggplot2 object.}

\item{n_samples}{The number of samples that will be displayed in the plot.}

\item{orientation}{The orientation of the plot ("horizontal"/"h" or "vertical"/"v").}

\item{factor}{A factor to adjust the dimensions of the plot.}
}
\value{
An updated version of the plot, with the appropriate orientation and
dimensions.
}
\description{
Process a ggplot2 bar plot to define its orientation and dimensions
}
