% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_barplot_ticknames}
\alias{process_barplot_ticknames}
\title{Process a ggplot2 bar plot to define its axis tick names}
\usage{
process_barplot_ticknames(plot, orientation, include_sample_names)
}
\arguments{
\item{plot}{A ggplot2 object.}

\item{orientation}{The orientation of the plot ("horizontal"/"h" or "vertical"/"v").}

\item{include_sample_names}{Whether sample names should be included in the plot.}
}
\value{
An updated version of the plot, with the appropriate tick names.
}
\description{
Process a ggplot2 bar plot to define its axis tick names
}
