% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_analysis}
\alias{run_analysis}
\title{Run an end-to-end SPARKI analysis}
\usage{
run_analysis(
  std_reports_path,
  mpa_reports_path,
  organism,
  reference_path,
  metadata_path,
  metadata_sample_col,
  metadata_columns,
  outdir_path,
  prefix,
  include_eukaryotes,
  include_sample_names,
  domain,
  samples_to_remove_path
)
}
\arguments{
\item{std_reports_path}{Path to a directory containing standard reports.}

\item{mpa_reports_path}{Path to a directory containing MPA-style reports.}

\item{organism}{Species-level name of the organism being analysed.}

\item{reference_path}{Path to 'inspect.txt' file in a Kraken2 reference database.}

\item{metadata_path}{Path to metadata table file.}

\item{metadata_sample_col}{Name of column containing sample IDs in the metadata table.}

\item{metadata_columns}{Name of columns of interest in the metadata table.}

\item{outdir_path}{Path to output directory.}

\item{prefix}{Prefix to be added to output file names.}

\item{include_eukaryotes}{Whether eukaryotes should be included in all plots (they are
included in some of the plots regardless of this flag).}

\item{include_sample_names}{Whether sample names should be included in all plots (they
are included in some of the plots regardless of this flag).}

\item{domain}{Domain(s) of interest.}

\item{samples_to_remove_path}{Path to file containing a list of samples that should
be removed from the SPARKI analysis.}
}
\value{
Processed reports ready for downstream analysis.
}
\description{
Run an end-to-end SPARKI analysis
}
