% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_sparki}
\alias{run_sparki}
\title{RUN AN END-TO-END SPARKI ANALYSIS}
\usage{
run_sparki(
  std_reports_path,
  mpa_reports_path,
  organism,
  reference_path,
  metadata_path,
  metadata_sample_col,
  metadata_columns,
  outdir_path,
  prefix,
  verbose,
  include_eukaryotes,
  include_sample_names,
  domain,
  samples_to_remove
)
}
\arguments{
\item{std_reports_path}{Path to a directory containing standard Kraken2 reports.}

\item{mpa_reports_path}{Path to a directory containing MPA-style reports.}

\item{organism}{Species being analysed (e.g. Homo sapiens).}

\item{reference_path}{Path to an "inspect.txt" file in a Kraken2 reference database.}

\item{metadata_path}{Path to metadata file.}

\item{metadata_sample_col}{The name of the column that contains sample IDs in the metadata table
specified via 'metadata_path'.}

\item{metadata_columns}{Comma-separated column names from the metadata table specified via 'metadata_path'.}

\item{outdir_path}{Path to an output directory.}

\item{prefix}{Prefix to be added to output files.}

\item{verbose}{Whether to run in verbose mode.}

\item{include_eukaryotes}{Whether eukaryotes should be included in all output plots (some plots will have
eukaryotes included regardless of this option).}

\item{include_sample_names}{Whether sample names should be included in all output plots (some plots will
have sample names included regardless of this option).}

\item{domain}{Domain(s) of interest (one of "Eukaryota", "Bacteria", "Viruses" or "Archaea").}

\item{samples_to_remove}{A list of samples that should not be included in the SPARKI analysis.}
}
\value{
Processed reports ready for downstream analysis.
}
\description{
RUN AN END-TO-END SPARKI ANALYSIS
}
