% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-subset.R
\name{subsetReports}
\alias{subsetReports}
\title{Subset report in order to filter out a given species.}
\usage{
subsetReports(report, species)
}
\description{
Subset report in order to filter out a given species.
}
