% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-add-info.R
\name{addMetadata}
\alias{addMetadata}
\title{ADD COLUMNS FROM ONE DATAFRAME TO ANOTHER}
\usage{
addMetadata(report, metadata, metadata_sample_col, metadata_columns)
}
\arguments{
\item{categories}{Categories of interest from df2 that should be added to df1.}
}
\value{
An updated version of df1.
}
\description{
This function takes as input a dataframe (df1) and adds metadata to it from another dataframe (df2). Both
dataframes must have the same sample IDs present, as these IDs will be used as an "anchor" between the tables.
Sample IDs can be specified as either row names or a standard column in the dataframes. The columns from df2
that should be added to df1 are specified by the user.
}
\details{
@param df1 Dataframe 1.
 @param df1_sample_col Name of column in df1 that contains sample IDs; alternatively, this can be "rownames".
 @param df2 Dataframe 2.
 @param df2_sample_col Name of column in df2 that contains sample IDs; alternatively, this can be "rownames".
}
