% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-add-info.R
\name{addSampleSize}
\alias{addSampleSize}
\title{Calculate sample sizes}
\usage{
addSampleSize(report)
}
\arguments{
\item{report}{A report generated by mergeReports() or loaded with load_STDreports().}
}
\value{
An updated version of the report provided as input, with a new column containing
sample sizes
}
\description{
Calculate sample sizes
}
