% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-assess-statistics.R
\name{assessStatistics}
\alias{assessStatistics}
\title{ASSESS THE STATISTICAL SIGNIFICANCE OF RESULTS}
\usage{
assessStatistics(report, ref_db, verbose)
}
\value{
An updated version of the input dataframe, with new columns containing statistical significance results.
}
\description{
ASSESS THE STATISTICAL SIGNIFICANCE OF RESULTS
}
