% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{check_directory}
\alias{check_directory}
\title{Carry out checks on a given directory}
\usage{
check_directory(dirpath, expectation)
}
\arguments{
\item{dirpath}{Path to a given directory.}

\item{expectation}{One of "empty" or "not_empty".}
}
\value{
Checked path to the given directory.
}
\description{
This function takes the path to a given directory and performs a few checks:
\itemize{
\item It checks if the directory exists.
\item It checks if the directory is empty or not depending on the expectation.
\item It ensures that the last character in the path is a slash symbol ("/").
}
}
\examples{
check_directory(dirpath = "path/to/directory", expectation = "empty")

}
\seealso{
\code{\link[=check_report_directory]{check_report_directory()}} for another function that carries out
checks based on a given directory path.
}
