% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_pdf_width}
\alias{determine_pdf_width}
\title{DETERMINE WIDTH FOR PDF FILE}
\usage{
determine_pdf_width(n_elements, base_size = 1, factor = 1)
}
\arguments{
\item{base_size}{Base plot width when there is only one element on the x-axis.}

\item{factor}{Value to help adjust the width.}
}
\value{
Plot width for PDF file.
}
\description{
@param n_elements Number of elements on the x-axis.
}
