% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_data}
\alias{load_data}
\title{Load and prepare data for SPARKI analysis}
\usage{
load_data(
  std_reports_path,
  mpa_reports_path,
  organism,
  reference_path,
  metadata_path,
  metadata_sample_col,
  metadata_columns,
  outdir_path,
  prefix,
  domain,
  samples_to_remove_path
)
}
\arguments{
\item{std_reports_path}{Path to a directory containing standard reports.}

\item{mpa_reports_path}{Path to a directory containing MPA-style reports.}

\item{organism}{Species-level name of the organism being analysed.}

\item{reference_path}{Path to 'inspect.txt' file in a Kraken2 reference database.}

\item{metadata_path}{Path to metadata table file.}

\item{metadata_sample_col}{Name of column containing sample IDs in the metadata table.}

\item{metadata_columns}{Name of columns of interest in the metadata table.}

\item{outdir_path}{Path to output directory.}

\item{prefix}{Prefix to be added to output file names.}

\item{domain}{Domain(s) of interest.}

\item{samples_to_remove_path}{Path to file containing a list of samples that should
be removed from the SPARKI analysis.}
}
\value{
A list containing the input arguments ready for the SPARKI analysis.
}
\description{
Load and prepare data for SPARKI analysis
}
