% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_STDreports}
\alias{load_STDreports}
\title{LOAD STANDARD REPORTS}
\usage{
load_STDreports(std_reports_dir, verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to output log messages.}

\item{mpa_reports_dir}{Path to a directory containing standard reports.}
}
\value{
A dataframe (tibble) with the content of all standard reports from the specified directory.
}
\description{
This function takes a path to a directory containing standard reports, reading and processing
all reports into a single dataframe.
}
