% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_delimited_list}
\alias{parse_delimited_list}
\title{Parse a symbol-delimited list to get individual elements}
\usage{
parse_delimited_list(del_list, delimiter)
}
\arguments{
\item{del_list}{A list with symbol-delimited values; the symbol can be a comma, for example.}

\item{delimiter}{The symbol that delimits the elements in the list.}
}
\value{
A vector with individual elements.
}
\description{
This function takes a symbol-delimited list of elements and splits it up into individual elements.
Symbols can be anything, e.g. ",", ";", "/", "//", "@", "." etc.
}
