% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-classification-summary.R
\name{plotClassificationSummary_violin}
\alias{plotClassificationSummary_violin}
\title{Plot the numbers of classified/unclassified reads per sample (violin plot)}
\usage{
plotClassificationSummary_violin(report, return_plot, outdir, prefix)
}
\arguments{
\item{report}{A report generated by mergeReports().}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}

\item{prefix}{The prefix that should be added to the output file name.}
}
\value{
If return_plot = TRUE, returns a violin plot with the classification summary.
}
\description{
This function takes a report and creates a violin plot showing the number of classified and
unclassified reads per sample. Note that reports that were processed with subsetReports()
should not be used with this function!
}
