% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-domain-reads.R
\name{plotDomainReads_violin}
\alias{plotDomainReads_violin}
\title{PLOT NUMBERS OF CLASSIFIED READS PER DOMAIN (VIOLIN PLOT)}
\usage{
plotDomainReads_violin(
  report,
  include_eukaryotes = FALSE,
  return_plot,
  outdir,
  prefix = "",
  fig_width,
  fig_height
)
}
\arguments{
\item{report}{Kraken2 report, either in standard or MPA-style format.}

\item{include_eukaryotes}{Whether to include eukaryotes in the plot.}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}

\item{prefix}{Prefix to be added to output plot name.}
}
\description{
This function takes a Kraken2 report, either in standard or MPA-style format, and creates
a violin plot showing the number of classified reads per domain.
}
