% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-minimisers.R
\name{plotMinimisers_dotplot}
\alias{plotMinimisers_dotplot}
\title{PLOT PROPORTION OF MINIMISERS AND SIGNIFICANCE PER SAMPLE}
\usage{
plotMinimisers_dotplot(
  report,
  domain,
  fig_width,
  fig_height,
  return_plot,
  outdir,
  prefix = ""
)
}
\arguments{
\item{report}{Kraken2 report, either in standard or MPA-style format.}

\item{return_plot}{Whether plot should be returned.}

\item{outdir}{Output directory where the plot should be saved.}

\item{prefix}{Prefix to be added to output plot name.}

\item{include_sample_names}{Whether sample names should be displayed.}

\item{orientation}{Whether plot should be horizontally or vertically oriented.}

\item{include_eukaryotes}{Whether to include eukaryotes in the plot.}
}
\description{
This function takes a Kraken2 report, either in standard or MPA-style format, and creates
a bar plot showing the number of classified reads per domain.
}
