% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_sparki}
\alias{run_sparki}
\title{PROCESS DATA}
\usage{
run_sparki(
  std_reports_path,
  mpa_reports_path,
  organism,
  reference_path,
  metadata_path,
  metadata_sample_col,
  metadata_columns,
  outdir_path,
  prefix,
  verbose,
  include_eukaryotes,
  include_sample_names,
  domain,
  remove
)
}
\arguments{
\item{std_reports_path}{Path to directory containing standard Kraken2 reports.}

\item{mpa_reports_path}{Path to directory containing MPA-style reports.}

\item{reference_path}{Path to file containing information on a Kraken2 reference database.}

\item{metadata_path}{Path to metadata file.}

\item{verbose}{Verbosity level.}
}
\value{
Processed reports ready for downstream analysis.
}
\description{
PROCESS DATA
}
